+++ CHANGELOG.md
@@ -2,17 +2,35 @@
 
 ## [Unreleased]
 
+### Changed
+- Feature to folder can be mapped with file extension .feature-folder and .feature-to-folder
+- Feature to file can be mapped with file extension .feature-file and .feature-to-file
+
 ### Added
+- Added a View of the metrics to the project
+
+## [0.0.4] - 2024-03-06
+
+### Added
+
+- Extension points for metrics.
+- A metrics service to calculate the different metrics.
+- Supported metrics:
+  - LineCount 
+  - FeatureFileMapping 
+  - Scattering 
+  - Tangling
+- Badge to the discord community.
 
 ### Changed
 
-### Deprecated
-
-### Removed
+- Upgrade Gradle version from 8.4 to 8.6.
+- Upgrade gradle files from groovy to kotlin. 
+- Upgrade action scripts to new versions
 
 ### Fixed
 
-### Security
+- Contributor link to Kuzzi04
 
 ## [0.0.3] - 2023-12-21
 
@@ -62,7 +80,8 @@ First Release to marketplace
 
 - Updated intellij plugin version to ```1.15.0```
 
-[Unreleased]: https://github.com/isselab/HAnS/compare/v0.0.3...HEAD
+[Unreleased]: https://github.com/isselab/HAnS/compare/v0.0.4...HEAD
+[0.0.4]: https://github.com/isselab/HAnS/compare/v0.0.3...v0.0.4
 [0.0.3]: https://github.com/isselab/HAnS/compare/v0.0.2...v0.0.3
 [0.0.2]: https://github.com/isselab/HAnS/compare/v0.0.1...v0.0.2
 [0.0.1]: https://github.com/isselab/HAnS/commits/v0.0.1